third_party/WebKit/LayoutTests/svg/dom/transform-parser-expected.txt

CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix()".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…trix(0,0,0,0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix( 78)2\t8\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(\t(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(.)2\t9e0 795\t0,- …".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e,4(7- )+e(e8.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(74- 9.(+ e.0e635".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(-9,).8e7e4e9ee ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(49-,,4)30378.052".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e7863e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(\t0".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e72(\te\t,49.7.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e55, ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(263,ee1)74-\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(.64110,.+1.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(.59e3...725-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(6\t-)2).180".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(,e 579\t++.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(-+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(74e\t-\t6\t-)29\t7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(221e..+\t+77".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix())4e\t1e).-e8378.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(\t4+01 \te".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(-6++9++ 9.3)6438".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(-,15e)(-+,- 6".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(3e+)085932".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(\t0944.6-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(\t5".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(3e(-2e3e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(6\t5e 3e\t(.\t1 69\t16".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(1ee5 56,(8.5\te,3".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e-\t\t e-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix()e0e)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(231293.16e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(()2(37-888580(9e…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(18e805".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(0".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(63) 693(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(80,2)+ 4ee,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(5e+59e2e5".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(5\t3e,9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e7\t)((59\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(963.+8-e73-e\t\t+)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(71(1(+-3382.3".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix( 27(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(+".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(88".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(9 43,+6 87.)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(\t-\t .5997".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(5.9)557".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(+65)e)0,\t217".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(.2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(+\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e8+(++\te52,. 73".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(,0+\t.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "…9-62159544,8,\t3\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(0857) 5".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(6,71+5e )e ".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(2e5\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(1e2)7e()\t90".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(9.-)7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(8e759.6,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(2(65\t.59e0+8-5".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(.0712".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e97,e33-)6,(e0-1…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(. 3 6.e,)57e7.\t(…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(1-7.(e(8.e-)e.) 6…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(6786- +8912e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix()ee8),-886().9\t-…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(.,e5732.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(18-+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(3,(13e+711e9eee0.)".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(6\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix()\t+69".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(, 19e\t)2)+0-(9)2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(1e,--+810)5..e37".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(35 )+e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(362020()8347081e78".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix()+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e-.54(+006\t3141e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e)\t.1+9e.0)7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(61)2e.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(7075-5(e67)+(-7)7)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e  ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(88)(3e21".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(3e-003(4".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(682- \t19-0)4642".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(6 7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(.9e1.e6+,e0-\t20e\t)9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(e+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(96,e2\t0 1e2-5)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix( 8-7(3.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "matrix(".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrixx(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matri(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix\u0000"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "matrix(\u0000)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate()".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0,0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "rotate(8 78".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(2\t8\t2e.172\t(0 .)".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "rotate(\t9".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(0 795\t0,- eee,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate((7- ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(+e(e8.)74- 9.(+ ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(.0e635(-9,).8e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e4e9ee ".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(49-,,4)30378.052".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e7863e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "rotate(\t0".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e72(\te\t,49.7.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e55, ".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "rotate(263,ee1)74-\t".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "rotate(.64110,.+1.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(.59e3...725-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(6\t-)2).180".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(,e 579\t++.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(-+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(74e\t-\t6\t-)29\t7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(221e..+\t+77".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate())4e\t1e).-e8378.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(\t4+01 \te".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "rotate(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(-6++9++ 9.3)6438".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(-,15e)(-+,- 6".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(3e+)085932".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(\t0944.6-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "rotate(\t5".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(3e(-2e3e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "rotate(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(6\t5e 3e\t(.\t1 69\t16".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(1ee5 56,(8.5\te,3".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e-\t\t e-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate()e0e)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(231293.16e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(()2(37-888580(9e…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(18e805".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "rotate(0".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(63) 693("
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(80,2)+ 4ee,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(5e+59e2e5".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(5\t3e,9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e7\t)((59\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(963.+8-e73-e\t\t+)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(71(1(+-3382.3".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate( 27(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "rotate(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(+".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "rotate(88".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(9 43,+6 87.)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(\t-\t .5997".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(+65)e)0,\t217"
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "rotate(.2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(+\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e8+(++\te52,. 73".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(,0+\t.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…e(-59-62159544,8,\t3\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0857) 5"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(6,71+5e )e ".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "rotate(2e5\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(1e2)7e()\t90"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(9.-)7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(8e759.6,".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(2(65\t.59e0+8-5".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "rotate(.0712".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e97,e33-)6,(e0-1…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(. 3 6.e,)57e7.\t(…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(1-7.(e(8.e-)e.) 6…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(6786- +8912e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate()ee8),-886().9\t-…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(.,e5732.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(18-+".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "rotate(3,(13e+711e9eee0.)".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "rotate(6\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate()\t+69".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(, 19e\t)2)+0-(9)2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(1e,--+810)5..e37".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(35 )+e"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(362020()8347081e78".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate()+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e-.54(+006\t3141e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e)\t.1+9e.0)7".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(61)2e."
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(7075-5(e67)+(-7)7)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e  ".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(88)(3e21"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(3e-003(4".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(682- \t19-0)4642".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "rotate(6 7".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "rotate(.9e1.e6+,e0-\t20e\t)9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(e+".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "rotate(96,e2\t0 1e2-5)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate( 8-7(3.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "rotate(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "rotate(8 78".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotatee(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotat(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate\u0000"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "rotate(\u0000)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale()".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(2\t8\t2e.172\t(0 .)".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(\t9".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(0 795\t0,- eee,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale((7- ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(+e(e8.)74- 9.(+ ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(.0e635(-9,).8e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e4e9ee ".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(49-,,4)30378.052".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e7863e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(\t0".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e72(\te\t,49.7.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e55, ".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "scale(263,ee1)74-\t".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "scale(.64110,.+1.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(.59e3...725-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(6\t-)2).180".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(,e 579\t++.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(-+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(74e\t-\t6\t-)29\t7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(221e..+\t+77".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale())4e\t1e).-e8378.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(\t4+01 \te".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "scale(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(-6++9++ 9.3)6438".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(-,15e)(-+,- 6".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(3e+)085932".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(\t0944.6-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(\t5".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(3e(-2e3e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "scale(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(6\t5e 3e\t(.\t1 69\t16".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(1ee5 56,(8.5\te,3".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e-\t\t e-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale()e0e)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(231293.16e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(()2(37-888580(9e…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(18e805".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(0".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(63) 693("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(80,2)+ 4ee,"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(5e+59e2e5".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(5\t3e,9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e7\t)((59\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(963.+8-e73-e\t\t+)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(71(1(+-3382.3".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale( 27(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "scale(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(+".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(88".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(9 43,+6 87.)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(\t-\t .5997".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(+65)e)0,\t217"
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(.2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(+\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e8+(++\te52,. 73".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(,0+\t.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…ale(-59-62159544,8,\t3\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0857) 5"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(6,71+5e )e ".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(2e5\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(1e2)7e()\t90"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(9.-)7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(8e759.6,".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(2(65\t.59e0+8-5".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(.0712".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e97,e33-)6,(e0-1…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(. 3 6.e,)57e7.\t(…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(1-7.(e(8.e-)e.) 6…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(6786- +8912e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale()ee8),-886().9\t-…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(.,e5732.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(18-+".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "scale(3,(13e+711e9eee0.)".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(6\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale()\t+69".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(, 19e\t)2)+0-(9)2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(1e,--+810)5..e37".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(35 )+e"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(362020()8347081e78".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale()+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e-.54(+006\t3141e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e)\t.1+9e.0)7".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(61)2e."
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(7075-5(e67)+(-7)7)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e  ".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(88)(3e21"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(3e-003(4".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(682- \t19-0)4642".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(6 7".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(.9e1.e6+,e0-\t20e\t)9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(e+".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "scale(96,e2\t0 1e2-5)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale( 8-7(3.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "scale(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "scale(8 78".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "scale(2\t8\t2e.172\t(0 .)".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scalee(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scal(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale\u0000"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "scale(\u0000)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX()".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(\t9".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(0 795\t0,- eee,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX((7- ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(+e(e8.)74- 9.(+ ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(.0e635(-9,).8e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e4e9ee ".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(49-,,4)30378.052".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e7863e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(\t0".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e72(\te\t,49.7.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e55, ".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(263,ee1)74-\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(.64110,.+1.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(.59e3...725-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(6\t-)2).180".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(,e 579\t++.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(-+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(74e\t-\t6\t-)29\t7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(221e..+\t+77".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX())4e\t1e).-e8378.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(\t4+01 \te".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "skewX(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(-6++9++ 9.3)6438".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(-,15e)(-+,- 6".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(3e+)085932".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(\t0944.6-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(\t5".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(3e(-2e3e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "skewX(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(6\t5e 3e\t(.\t1 69\t16".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(1ee5 56,(8.5\te,3".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e-\t\t e-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX()e0e)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(231293.16e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(()2(37-888580(9e…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(18e805".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(0".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(63) 693("
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(80,2)+ 4ee,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(5e+59e2e5".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(5\t3e,9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e7\t)((59\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(963.+8-e73-e\t\t+)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(71(1(+-3382.3".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX( 27(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "skewX(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(+".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(88".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(9 43,+6 87.)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(\t-\t .5997".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(+65)e)0,\t217"
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(.2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(+\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e8+(++\te52,. 73".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(,0+\t.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(-59-62159544,8,\t3\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0857) 5"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(6,71+5e )e ".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(2e5\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(1e2)7e()\t90"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(9.-)7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(8e759.6,".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(2(65\t.59e0+8-5".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(.0712".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e97,e33-)6,(e0-1…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(. 3 6.e,)57e7.\t(…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(1-7.(e(8.e-)e.) 6…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(6786- +8912e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX()ee8),-886().9\t-…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(.,e5732.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(18-+".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(3,(13e+711e9eee0.…".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(6\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX()\t+69".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(, 19e\t)2)+0-(9)2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(1e,--+810)5..e37".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(35 )+e"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(362020()8347081e78".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX()+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e-.54(+006\t3141e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e)\t.1+9e.0)7".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(61)2e."
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(7075-5(e67)+(-7)7)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e  ".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(88)(3e21"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(3e-003(4".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(682- \t19-0)4642".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(6 7".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(.9e1.e6+,e0-\t20e\t)9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(e+".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(96,e2\t0 1e2-5)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX( 8-7(3.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "skewX(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(8 78".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewX(2\t8\t2e.172\t(0 .)".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewX(\t9".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewXX(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skew(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX\u0000"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewX(\u0000)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY()".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0 795\t0,- eee,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY((7- ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(+e(e8.)74- 9.(+ ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(.0e635(-9,).8e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e4e9ee ".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(49-,,4)30378.052".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e7863e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewY(\t0".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e72(\te\t,49.7.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e55, ".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(263,ee1)74-\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(.64110,.+1.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(.59e3...725-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(6\t-)2).180".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(,e 579\t++.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(-+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(74e\t-\t6\t-)29\t7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(221e..+\t+77".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY())4e\t1e).-e8378.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(\t4+01 \te".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "skewY(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(-6++9++ 9.3)6438".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(-,15e)(-+,- 6".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(3e+)085932".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(\t0944.6-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewY(\t5".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(3e(-2e3e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "skewY(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(6\t5e 3e\t(.\t1 69\t16".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(1ee5 56,(8.5\te,3".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e-\t\t e-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY()e0e)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(231293.16e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(()2(37-888580(9e…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(18e805".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewY(0".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(63) 693("
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(80,2)+ 4ee,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(5e+59e2e5".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(5\t3e,9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e7\t)((59\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(963.+8-e73-e\t\t+)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(71(1(+-3382.3".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY( 27(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "skewY(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(+".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewY(88".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(9 43,+6 87.)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(\t-\t .5997".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(+65)e)0,\t217"
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewY(.2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(+\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e8+(++\te52,. 73".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(,0+\t.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(-59-62159544,8,\t3\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0857) 5"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(6,71+5e )e ".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewY(2e5\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(1e2)7e()\t90"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(9.-)7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(8e759.6,".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(2(65\t.59e0+8-5".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewY(.0712".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e97,e33-)6,(e0-1…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(. 3 6.e,)57e7.\t(…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(1-7.(e(8.e-)e.) 6…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(6786- +8912e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY()ee8),-886().9\t-…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(.,e5732.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(18-+".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(3,(13e+711e9eee0.…".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewY(6\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY()\t+69".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(, 19e\t)2)+0-(9)2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(1e,--+810)5..e37".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(35 )+e"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(362020()8347081e78".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY()+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e-.54(+006\t3141e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e)\t.1+9e.0)7".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(61)2e."
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(7075-5(e67)+(-7)7)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e  ".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(88)(3e21"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(3e-003(4".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(682- \t19-0)4642".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(6 7".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(.9e1.e6+,e0-\t20e\t)9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(e+".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(96,e2\t0 1e2-5)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY( 8-7(3.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "skewY(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(8 78".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(2\t8\t2e.172\t(0 .)".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "skewY(\t9".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "skewY(0 795\t0,- eee,".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewYY(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skew(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY\u0000"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "skewY(\u0000)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate()".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0,0,0,0,0,0,0,0,0,0…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate((7- ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(+e(e8.)74- 9.(+ ".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(.0e635(-9,).8e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e4e9ee ".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(49-,,4)30378.052".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e7863e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(\t0".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e72(\te\t,49.7.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e55, ".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "translate(263,ee1)74-\t".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "…ranslate(.64110,.+1.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(.59e3...725-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(6\t-)2).180".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(,e 579\t++.".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(-+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(74e\t-\t6\t-)29\t7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(221e..+\t+77".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate())4e\t1e).-e8378.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…ranslate(\t4+01 \te".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "translate(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(-6++9++ 9.3)6438".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(-,15e)(-+,- 6".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(3e+)085932".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(-".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…ranslate(\t0944.6-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(\t5".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(3e(-2e3e".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "translate(".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(6\t5e 3e\t(.\t1 69\t16".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(1ee5 56,(8.5\te,3".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e-\t\t e-".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate()e0e)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…nslate(231293.16e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(()2(37-888580(9e…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(18e805".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(0".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(63) 693("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(80,2)+ 4ee,"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(5e+59e2e5".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(5\t3e,9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e7\t)((59\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(963.+8-e73-e\t\t+)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(71(1(+-3382.3".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate( 27(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "translate(".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(+".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(88".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(9 43,+6 87.)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(\t-\t .5997".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(+65)e)0,\t217"
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(.2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(+\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e8+(++\te52,. 73".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(,0+\t.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "…ate(-59-62159544,8,\t3\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0857) 5"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(6,71+5e )e ".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(2e5\t".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(1e2)7e()\t90"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(9.-)7".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(8e759.6,".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(2(65\t.59e0+8-5".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(.0712".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e97,e33-)6,(e0-1…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(. 3 6.e,)57e7.\t(…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(1-7.(e(8.e-)e.) 6…".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(6786- +8912e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate()ee8),-886().9\t-…".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(.,e5732.".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(18-+".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "translate(3,(13e+711e9eee0.)".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(6\t".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate()\t+69".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(, 19e\t)2)+0-(9)2".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(1e,--+810)5..e37".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(35 )+e"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(362020()8347081e78".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate()+".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e-.54(+006\t3141e".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e)\t.1+9e.0)7".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(61)2e."
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(7075-5(e67)+(-7)7)".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e  ".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(88)(3e21"
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(3e-003(4".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(682- \t19-0)4642".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(6 7".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(.9e1.e6+,e0-\t20e\t)9".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(e+".
CONSOLE ERROR: Error: <g> attribute transform: Trailing garbage, "translate(96,e2\t0 1e2-5)".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate( 8-7(3.".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected number, "translate(".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(8 78".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(2\t8\t2e.172\t(0 .)".
CONSOLE ERROR: Error: <g> attribute transform: Unexpected end of attribute. Expected ')', "translate(\t9".
CONSOLE ERROR: Error: <g> attribute transform: Expected ')', "translate(0 795\t0,- eee,".
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate((7- ".
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translatee(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translat(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate\u0000"
CONSOLE ERROR: Error: <g> attribute transform: Expected number, "translate(\u0000)".
This test fuzzes the transform parser with semi-random attribute values and dumps the results of any values that parse successfully.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Parsed as 1 transform(s): matrix(0,0,0,0,0,0)
Parsed as 1 transform(s): rotate(0)
Parsed as 1 transform(s): rotate(0,0,0)
Parsed as 1 transform(s): scale(0)
Parsed as 1 transform(s): scale(0,0)
Parsed as 1 transform(s): skewX(0)
Parsed as 1 transform(s): skewY(0)
Parsed as 1 transform(s): translate(0)
Parsed as 1 transform(s): translate(0,0)
PASS successfullyParsed is true

TEST COMPLETE
